% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compileSettings}
\alias{compileSettings}
\title{compiles a valid XML distance description into a handle}
\usage{
compileSettings(setting)
}
\arguments{
\item{setting}{an XML string describing the distance to be used}

\item{setting}{an XML string describing the distance to be used}
}
\value{
An integer handle to pass the (compiled) XML settings to other functions
}
\description{
compiles a valid XML distance description into a handle
}
\details{
Distance functions are described via a specific XML format. Consider reading 
the documentation at \url{trajectorycomputing.com/libtrajcomp-xml}
}

