% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martin.R
\name{distance_vector}
\alias{distance_vector}
\title{Calculates a distance vector for a distance given by name}
\usage{
distance_vector(dataset, query, distance_name)
}
\arguments{
\item{dataset}{A matrix (or data frame) representing the dataset with NaNs between different trajectories}

\item{query}{A matrix (or data frame) representing a single query trajectory}

\item{distance_name}{A string defining the distance by name}
}
\value{
A vector containing the distance of each trajectory as a column in the database (see the example)
}
\description{
Calculates a distance vector for a distance given by name
}
\examples{
distance_vector(prague,tsplit(prague)[[1]],"closest_pair");
}

