% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{getTrajectoryIDs}
\alias{getTrajectoryIDs}
\title{Create a vector with an integer trajectory ID from a NaN separated 
dataset}
\usage{
getTrajectoryIDs(Q)
}
\arguments{
\item{x}{A trajectory database separated by NaN lines}
}
\value{
A vector containing trajectory indizes and NaN
}
\description{
Create a vector with an integer trajectory ID from a NaN separated 
dataset
}
\examples{
prague$tid = getTrajectoryIDs(prague);
}
\seealso{
\code{\link{removeNaNs}} to create a dataset with an ID 
column and without NaNs (ideal for \code{\link{ddply}})

Other basic trajectory dataset tools: \code{\link{removeNaNs}},
  \code{\link{tsplit}}
}

