% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff2d.R
\name{homogene}
\alias{homogene}
\title{Move a coordinate from projective space into Euclidean space}
\usage{
homogene(Q)
}
\arguments{
\item{Q}{A vector or matrix}
}
\value{
A vector with the last column removed
}
\description{
Move a coordinate from projective space into Euclidean space
}
\seealso{
Other affine: \code{\link{PM.affine}},
  \code{\link{PM.translation}}, \code{\link{map_affine}},
  \code{\link{map_scale}}, \code{\link{map_skew}},
  \code{\link{map_translate}},
  \code{\link{projective.example}},
  \code{\link{projective}}
}

