% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff2d.R
\name{map_affine}
\alias{map_affine}
\title{Applies an affine map given by params}
\usage{
map_affine(Q, params)
}
\arguments{
\item{Q}{A vector or matrix}

\item{params}{A vector containing 6 parameters defining the affine mapping}
}
\value{
A vector or matrix, after applying the transformation in projective space
}
\description{
Applies an affine map given by params
}
\seealso{
Other affine: \code{\link{PM.affine}},
  \code{\link{PM.translation}}, \code{\link{homogene}},
  \code{\link{map_scale}}, \code{\link{map_skew}},
  \code{\link{map_translate}},
  \code{\link{projective.example}},
  \code{\link{projective}}
}

