% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff2d.R
\name{map_skew}
\alias{map_skew}
\title{Applies a skew translation as given by a andbt}
\usage{
map_skew(Q, a, b)
}
\arguments{
\item{Q}{A vector or matrix}

\item{a}{Parameter in first direction}

\item{b}{Parameter in second direction}
}
\value{
A vector or matrix, after applying the transformation in projective space
}
\description{
Applies a skew translation as given by a andbt
}
\seealso{
Other affine: \code{\link{PM.affine}},
  \code{\link{PM.translation}}, \code{\link{homogene}},
  \code{\link{map_affine}}, \code{\link{map_scale}},
  \code{\link{map_translate}},
  \code{\link{projective.example}},
  \code{\link{projective}}
}

