% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff2d.R
\name{map_translate}
\alias{map_translate}
\title{Applies a translation as given by s and t}
\usage{
map_translate(Q, s, t)
}
\arguments{
\item{Q}{A vector or matrix}

\item{s}{Parameter in first direction}

\item{t}{Parameter in second direction}
}
\value{
A vector or matrix, after applying the transformation in projective space
}
\description{
Applies a translation as given by s and t
}
\seealso{
Other affine: \code{\link{PM.affine}},
  \code{\link{PM.translation}}, \code{\link{homogene}},
  \code{\link{map_affine}}, \code{\link{map_scale}},
  \code{\link{map_skew}}, \code{\link{projective.example}},
  \code{\link{projective}}
}

