% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_distance.R
\name{matrixcross}
\alias{matrixcross}
\title{Calculates the cross product sparse matrix from two incidence matrices}
\usage{
matrixcross(A, B)
}
\arguments{
\item{Matrix}{A: A one-sparse matrix (class matrix, not dgCMatrix, so not sparsely stored) with rownames}

\item{Matrix}{B: A second one-sparse matrix (class matrix, not dgCMatrix, so not sparsely stored) with rownames}
}
\value{
A matrix of nrow(A) * nrow(B) rows, with rownames concatenated and combinations set to one (e.g., the cross product of the alphabets)
}
\description{
Calculates the cross product sparse matrix from two incidence matrices
}
\examples{
matrixcross(A,B)
}
\seealso{
Other SVDdistance: \code{\link{matrixcross_nonames}}
}

