% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_distance.R
\name{matrixcross_nonames}
\alias{matrixcross_nonames}
\title{Calculates the cross product sparse matrix from two incidence matrices (faster than the version with rownames)}
\usage{
matrixcross_nonames(A, B)
}
\arguments{
\item{Matrix}{A: A one-sparse matrix (class matrix, not dgCMatrix, so not sparsely stored) with or without rownames}

\item{Matrix}{B: A second one-sparse matrix (class matrix, not dgCMatrix, so not sparsely stored) with or without rownames}
}
\value{
A matrix of nrow(A) * nrow(B) rows, without rownames concatenated and combinations set to one (e.g., the cross product of the alphabets)
}
\description{
Calculates the cross product sparse matrix from two incidence matrices (faster than the version with rownames)
}
\examples{
matrixcross_nonames(A,B)
}
\seealso{
Other SVDdistance: \code{\link{matrixcross}}
}

