% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff2d.R
\name{PM.affine}
\alias{PM.affine}
\title{Creates a general affine mapping in projective space}
\usage{
PM.affine(params)
}
\arguments{
\item{params}{A vector containing 6 parameters defining the affine mapping}
}
\value{
A 3x3 matrix in projective space
}
\description{
Creates a general affine mapping in projective space
}
\seealso{
Other affine: \code{\link{PM.translation}},
  \code{\link{homogene}}, \code{\link{map_affine}},
  \code{\link{map_scale}}, \code{\link{map_skew}},
  \code{\link{map_translate}},
  \code{\link{projective.example}},
  \code{\link{projective}}
}

