% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff2d.R
\name{PM.translation}
\alias{PM.translation}
\title{Creates a translation matrix projective space}
\usage{
PM.translation(s, t)
}
\arguments{
\item{s}{Movement in first direction}

\item{t}{Movement in second direction}
}
\value{
A 3x3 matrix in projective space
}
\description{
Creates a translation matrix projective space
}
\seealso{
Other affine: \code{\link{PM.affine}},
  \code{\link{homogene}}, \code{\link{map_affine}},
  \code{\link{map_scale}}, \code{\link{map_skew}},
  \code{\link{map_translate}},
  \code{\link{projective.example}},
  \code{\link{projective}}
}

