% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff2d.R
\name{projective.example}
\alias{projective.example}
\title{Shows, how to use the map_affine functinos to randomly project a triangle}
\usage{
projective.example()
}
\description{
Shows, how to use the map_affine functinos to randomly project a triangle
}
\seealso{
Other affine: \code{\link{PM.affine}},
  \code{\link{PM.translation}}, \code{\link{homogene}},
  \code{\link{map_affine}}, \code{\link{map_scale}},
  \code{\link{map_skew}}, \code{\link{map_translate}},
  \code{\link{projective}}
}

