% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff2d.R
\name{projective}
\alias{projective}
\title{Move a coordinate into its projective space}
\usage{
projective(Q)
}
\arguments{
\item{Q}{A vector or matrix}
}
\value{
A vector with a column 1 added
}
\description{
Move a coordinate into its projective space
}
\seealso{
Other affine: \code{\link{PM.affine}},
  \code{\link{PM.translation}}, \code{\link{homogene}},
  \code{\link{map_affine}}, \code{\link{map_scale}},
  \code{\link{map_skew}}, \code{\link{map_translate}},
  \code{\link{projective.example}}
}

