% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{tsplit}
\alias{tsplit}
\title{Split trajectory dataset into a list}
\usage{
tsplit(x)
}
\arguments{
\item{x}{A trajectory database separated by NaN lines}
}
\value{
A list of trajectories
}
\description{
Split trajectory dataset into a list
}
\examples{
psplit = tsplit(prague);
}
\seealso{
\code{\link{getTrajectoryIDs}} and \code{\link{ddply}}

Other basic trajectory dataset tools: \code{\link{getTrajectoryIDs}},
  \code{\link{removeNaNs}}
}

